int to size_t

+++ src/tests/c_api/test_join_sail.cpp
@@ -461,7 +461,7 @@ int main(int argc, char** argv) {
         }
 
         // Print P1 header
-        for (int i = 1; i< js1_header.size(); i++){
+        for (size_t i = 1; i< js1_header.size(); i++){
             int curr = js1_header[i];
             std::cout << curr;
             if (i != js1_header.size()-1){
